pyproject.toml
[project]
name = "glossaryapi"
version = "0.1.0"
description = "Glossary API with FastAPI, SQLite, static OpenAPI docs, Dockerized"
authors = [
  { name = "User" }
]
requires-python = ">=3.10"
readme = "README.md"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "sqlmodel>=0.0.21",
  "sqlalchemy>=2.0.32",
  "grpcio>=1.60.0",
  "grpcio-tools>=1.60.0",
  "protobuf>=4.25.0"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "anyio>=4.0.0",
  "locust>=2.20.0"
]

[tool.uv]
managed = true

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling>=1.22.2"]
build-backend = "hatchling.build"
